+++ Enums/JuiceEffect.cs
@@ -1,0 +1,15 @@
+﻿namespace JuicesMod.Enums
+{
+    public enum JuiceEffect
+    {
+        None = 0x0,
+        Healing = 0x1,
+        Damages = 0x2,
+        Marathoner = 0x4,
+        Asmathic = 0x8,
+        NightVision = 0x10,
+        UpsideDown = 0x20,
+        EMP = 0x40,
+        SignalJamming = 0x80,
+    }
+}